added password hashing
