+++ src/main/java/com/ncedu/cheetahtest/security/service/AuthService.java
@@ -1,5 +1,7 @@
 package com.ncedu.cheetahtest.security.service;
 
+import com.ncedu.cheetahtest.developer.entity.ResetToken;
+import com.ncedu.cheetahtest.mail.entity.PasswordDTO;
 import com.ncedu.cheetahtest.security.entity.AccessTokenDto;
 import com.ncedu.cheetahtest.security.entity.LoginDto;
 import com.ncedu.cheetahtest.security.entity.RegisterDto;
@@ -7,4 +9,7 @@ import com.ncedu.cheetahtest.security.entity.RegisterDto;
 public interface AuthService {
     void register(RegisterDto registerDto);
     AccessTokenDto login(LoginDto loginDto);
+
+    void changeUserPassword(ResetToken resetToken, String password);
+    boolean validatePassword(PasswordDTO passwordDTO);
 }
